XbOne BT fix
Xbox One Wireless controllers in some circumstances fail the Windows XInput filter.

The old firmware presents a good state that will fall through to recieve a map via SDL db, but new firmware (in a bad state) presents no trigger ouput and a non-normative mapping.

This patch adds accommodations to make the best of new firmware when it's in a bad state. This is an XInput related issue and affects Windows only.