Address Mac Steam Input Update
Steam is now spoofing an Xbox controller on MacOS (as on Windows) instead of the former cleaner, dedicated, "Steam Virtual Controller" device (as used on Linux). This appears to be a hasty/broken change because the virtual devices frequently show up in duplicate, and Steam itself seems to fail at recognizing them as virtualized, enumerating them as additional mappable devices.

Fortunately (???) Apple has killed connectivity support for the real device being spoofed as of BigSur (Xbox 360 wired controller), so we can discard it outright via blacklist.